+++ src/Core.php
@@ -10,6 +10,8 @@
 
 namespace Fragen\Rollback;
 
+use stdClass;
+
 /*
  * Exit if called directly.
  */
@@ -70,6 +72,7 @@ class Core {
 	public function load_hooks() {
 		add_filter( 'http_request_args', [ $this, 'filter_http_request_args' ], 10, 2 );
 		add_filter( 'pre_http_request', [ $this, 'filter_http_request' ], 10, 3 );
+		add_filter( 'site_transient_update_core', [ $this, 'add_rollback_offer' ], 10, 1 );
 	}
 
 	/**
@@ -100,10 +103,10 @@ class Core {
 	 * @param array  $args   Array of filter args.
 	 * @param string $url    URL from filter.
 	 *
-	 * @return \stdClass $response Output from wp_remote_get().
+	 * @return stdClass $response Output from wp_remote_get().
 	 */
 	public function filter_http_request( $result, $args, $url ) {
-		if ( $result || isset( $args['_core_rollback'] ) ) {
+		if ( $result || isset( $args['_core_rollback'] ) || ! isset( $args['_rollback_version'] ) ) {
 			return $result;
 		}
 		if ( false === strpos( $url, '//api.wordpress.org/core/version-check/' ) ) {
@@ -128,17 +131,17 @@ class Core {
 	/**
 	 * Update Core API update response to add rollback.
 	 *
-	 * @param \stdClass $response         Core API response.
-	 * @param string    $rollback_version Rollback version.
+	 * @param stdClass $response         Core API response.
+	 * @param string   $rollback_version Rollback version.
 	 *
-	 * @return \stdClass
+	 * @return stdClass
 	 */
 	public function set_rollback( $response, $rollback_version ) {
 		if ( ! array_key_exists( $rollback_version, static::$core_versions ) ) {
 			return $response;
 		}
 		$body   = wp_remote_retrieve_body( $response );
-		$offers = \json_decode( $body );
+		$offers = json_decode( $body );
 		$latest = false;
 		$num    = count( array_keys( $offers->offers ) );
 		foreach ( array_keys( $offers->offers ) as $key ) {
@@ -155,10 +158,33 @@ class Core {
 		if ( ! $latest && ( ( $key + 1 ) === $num ) ) {
 			$offers->offers[ $num ] = static::$core_versions[ $rollback_version ];
 		}
+		set_site_transient( '_core_rollback_offers', $offers->offers, 15 );
 
-		$body             = \json_encode( $offers );
+		$body             = json_encode( $offers );
 		$response['body'] = $body;
 
 		return $response;
+	}
+
+	/**
+	 * Add core rollback offer to update_core transient.
+	 *
+	 * @param stdClass $transient Update core transient.
+	 *
+	 * @return stdClass
+	 */
+	public function add_rollback_offer( $transient ) {
+		$rollback = get_site_transient( '_core_rollback' );
+		$version  = $rollback['core_dropdown'] ?? '';
+		$offers   = get_site_transient( '_core_rollback_offers' );
+		$versions = $this->get_core_versions();
+		if ( array_keys( $versions )[0] === $version ) {
+			unset( $transient->updates[1] );
+		}
+		if ( $offers && 'latest' === $offers[1]->response && $version === $offers[1]->version ) {
+			$transient->updates[] = $offers[1];
+		}
+
+		return $transient;
 	}
 }